+++ APIJava/src/main/java/com/farmacia/sanrafael/APIJava/entities/VentaEntity.java
@@ -14,7 +14,7 @@ public class VentaEntity {
     @Id
     @Column(columnDefinition = "INT")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_venta;
+    private Integer id_venta;
 
     @Temporal(TemporalType.TIMESTAMP)
     @Column(columnDefinition = "DATETIME", nullable = false)
